fix(adapter-next): prevent `revalidatePath` error in the slice simulator
