+++ packages/adapter-next/src/simulator/index.ts
@@ -4,5 +4,5 @@ export type {
 	SliceSimulatorSliceZoneProps,
 } from "./SliceSimulator";
 
-export { getSlices } from "./getSlices";
+export { getSlices } from "./react-server/getSlices";
 export type { SliceSimulatorParams } from "./types";
